ru.appline/src/main/java/task_5/logic/Calculator.java
package task_5.logic;

public class Calculator implements Add, Subtract, Multiply, Divide {
    private int a;
    private int b;
    private String operator;

    public Calculator() {
        super();
    }

    public Calculator(int a, int b, String operator) {
        this.a = a;
        this.b = b;
        this.operator = operator;
    }

    public int getA() {
        return a;
    }

    public void setA(int a) {
        this.a = a;
    }

    public int getB() {
        return b;
    }

    public void setB(int b) {
        this.b = b;
    }

    public String getOperator() {
        return operator;
    }

    public void setOperator(String operator) {
        this.operator = operator;
    }

    public int add(int a, int b) {
        return a + b;
    }

    public int subtract(int a, int b) {
        return a - b;
    }

    public int multiply(int a, int b) {
        return a * b;
    }

    public int divide(int a, int b) {
        return a / b;
    }
}
